FeatureComplete: check if target dir is a directory
This commit improves the way in which the FeatureComplete\Config class handles the target directories passed in the command-line. Before it would only check if the target dir exists. Now, it also checks if it is a directory.

This avoids the following error when passing a file to the script:

```
bin/phpcs-check-feature-completeness pathToFile
PHP Fatal error:  Uncaught UnexpectedValueException: RecursiveDirectoryIterator::__construct(pathToFile): Failed to open directory: Not a directory in Scripts/Utils/FileList.php:65
```

The script will now throw an uncaught exception. I considered handling the exception as out of the scope of this change, an uncaught exception is also thrown when an invalid path is passed.